Refactor: reorder Groupie methods by importance
I'm not sure what my original reasoning was, but it's nice to be able
read a class from the top to the bottom and see the most frequently
useful methods first.

In this case, classifying a text is more important than a classifying a
word, which seems like it should have been a private method. Ditto for
unique words, which might have some edge case usage as public method,
but only is used by one specific weighting strategy.