Refactor: reduce complexity of Groupie#unique_words
Another instance of applying 10 years of new core Ruby APIs to more
efficiently iterate over collections and sum numbers.

One sort-of-change is that I've removed memoization, which was a
performance boost for repeated classifications, but did not get reset
when new words got added to a group, so there was a subtle bug based on
how it was used. Now it might be (slightly?) slower, but it's not
bugged.