Change: Groups track their total word count
This is a refactoring with side effects, so it's a change.

Tracking this means we don't have to count it manually in
Groupie#default_weight. A side effect is that old serialized data won't
have this cache data, so its smart weight default weight will not be
correct. Adding old data to a new instance should fix this.