Feat: Groupie.tokenize breaks down urls more intelligently
Prior to this URLs were stripped of all special characters and ended up
as word garbage. Now splitting happens more intelligently: parse the
URL, use the parser to extract main segments, then split those based on
common patterns (such as using hyphens and underscores as stand-in for
spaces to split words, dealing with query strings and anchors.

There is room for improvement, but it's a first stab at improving the
tokenizer.