Refactor: Groupie.tokenize modifies a duplicate string
This should cut down the number of new Strings created by a large
amount. Previously we created at least 5 modified copies of the input
String, a copy of each part of a URI and a copy of each final word.
Now we copy the input once and modify the rest where possible.