Fix: don't raise on invalid URIs
In the initial implementation to support URIs, I overlooked that it
could raise an exception. This is now handled, in which case we treat
the input String as plain text, instead of decomposing it into tokens.

The code triggers Rubocop's desire to have shorter methods, so I'll see
about reducing the complexity of the method next.