Breaking: drop Ruby 2.6 support; minimum becomes 2.7.0
Ruby 2.6 went End of Life just over a month ago, so it's a good time to
drop support for it. The minimum supported version is now Ruby 2.7.0

If Ruby remains consistent, 2.7 will be EOL in March/April of 2023.

While doing this, updated CI to use 2.7 for non-test tools. This makes
Rubocop happier.